TrainerDex Rework: Explicitly set `db_table` on trainer models
This explicitly sets the `db_table` values so they match the names of
the existing database tables.

This will prevent django from using the new defaults which will change
once the app label does.

Once the default changes, django will think you're changing the table
name instead of just solidifying it. To work around this, we doctor the
history to make it look as though the table was created with this
"strange" table name.